вынесены domain и data слои
