+++ app/build.gradle
@@ -34,7 +34,9 @@ android {
 
 dependencies {
 
-//    implementation project(path: ':domain')
+    implementation project(path: ':domain')
+    implementation project(path: ':data')
+
     implementation 'androidx.core:core-ktx:1.12.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
     implementation 'com.google.android.material:material:1.9.0'
