--- app/src/main/java/com/asefactory/getapplicationlasttimeusage/domain/models/ApplicationInfoWithLastUsageTime.kt
@@ -1,8 +1,0 @@
-package com.asefactory.getapplicationlasttimeusage.domain.models
-
-data class ApplicationInfoWithLastUsageTime(
-    val icon: String,
-    val packageName: String,
-    val label: String,
-    val lastUsageTime: Long,
-)
