--- app/src/main/java/com/asefactory/getapplicationlasttimeusage/domain/models/ApplicationInfoWithUsageTimeByPeriod.kt
@@ -1,8 +1,0 @@
-package com.asefactory.getapplicationlasttimeusage.domain.models
-
-data class ApplicationInfoWithUsageTimeByPeriod(
-    val icon: String,
-    val packageName: String,
-    val label: String,
-    val usageTimeByPeriod: Long,
-)
