--- app/src/main/java/com/asefactory/getapplicationlasttimeusage/domain/repositories/AppsRepository.kt
@@ -1,7 +1,0 @@
-package com.asefactory.getapplicationlasttimeusage.domain.repositories
-
-import android.content.pm.PackageInfo
-
-interface AppsRepository {
-    fun getApplicationsList(): Set<PackageInfo>
-}
