--- app/src/main/java/com/asefactory/getapplicationlasttimeusage/domain/repositories/UsageInformationRepository.kt
@@ -1,7 +1,0 @@
-package com.asefactory.getapplicationlasttimeusage.domain.repositories
-
-import android.app.usage.UsageStats
-
-interface UsageInformationRepository {
-    fun getUsageInformationAboutApplications(): Set<UsageStats>
-}
