--- app/src/main/java/com/asefactory/getapplicationlasttimeusage/domain/usecases/GetApplicationLastUsingDateUseCase.kt
@@ -1,10 +1,0 @@
-package com.asefactory.getapplicationlasttimeusage.domain.usecases
-
-import com.asefactory.getapplicationlasttimeusage.domain.models.ApplicationInfoWithLastUsageTime
-
-class GetApplicationLastUsingDateUseCase {
-
-    fun execute(): List<ApplicationInfoWithLastUsageTime>{
-        return emptyList()
-    }
-}
