--- app/src/main/java/com/asefactory/getapplicationlasttimeusage/domain/usecases/GetApplicationUsageTimeByPeriodUseCase.kt
@@ -1,10 +1,0 @@
-package com.asefactory.getapplicationlasttimeusage.domain.usecases
-
-import com.asefactory.getapplicationlasttimeusage.domain.models.ApplicationInfoWithUsageTimeByPeriod
-
-class GetApplicationUsageTimeByPeriodUseCase {
-
-    fun execute(period: Int): List<ApplicationInfoWithUsageTimeByPeriod> {
-        return emptyList()
-    }
-}
