+++ app/src/main/java/com/asefactory/getapplicationlasttimeusage/ui/activity/MainActivity.kt
@@ -15,6 +15,10 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
+        setupMainScreen()
+    }
+
+    private fun setupMainScreen() {
         val navHostFragment =
             supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
         val navController = navHostFragment.navController
