+++ app/src/main/java/com/asefactory/getapplicationlasttimeusage/ui/fragments/installation_date/InstallationDatesFragment.kt
@@ -5,9 +5,9 @@ import android.view.View
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.asefactory.getapplicationlasttimeusage.domain.usecases.GetApplicationInstallTimeUseCase
+import com.asefactory.data.AppsListWithDateRepositoryImpl
+import com.asefactory.domain.usecases.GetApplicationInstallTimeUseCase
 import com.asefactory.getapplicationlasttimeusage.R
-import com.asefactory.getapplicationlasttimeusage.data.AppsListWithDateRepositoryImpl
 import com.asefactory.getapplicationlasttimeusage.ui.adapters.AppListAdapter
 
 class InstallationDatesFragment : Fragment(R.layout.fragment_installation_dates) {
