+++ domain/src/main/java/com/asefactory/domain/repositories/AppsRepository.kt
@@ -1,0 +1,7 @@
+package com.asefactory.domain.repositories
+
+import com.asefactory.domain.models.ApplicationInfoWithInstallationDate
+
+interface AppsRepository {
+    fun getApplicationsList(): List<ApplicationInfoWithInstallationDate>
+}
