+++ settings.gradle
@@ -3,8 +3,9 @@ dependencyResolutionManagement {
     repositories {
         google()
         mavenCentral()
-        jcenter() // Warning: this repository is going to shut down soon
     }
 }
 rootProject.name = "GetApplicationLastTimeUsage"
 include ':app'
+include ':domain'
+include ':data'
